--- __init__.py
@@ -1,7 +1,0 @@
-print ('LOAD __INIT__')
-GRID_DIR = 'grids'
-global GRID_DIR
-TEXT_DIR = 'text'
-global TEXT_DIR
-SOLUTION_DIR = 'solutions'
-global SOLUTION_DIR
