+++ bedlam/__init__.py
@@ -1,0 +1,10 @@
+GRID_DIR = 'grids'
+TEXT_DIR = 'text'
+SOLUTION_DIR = 'solutions'
+
+# import bedlam.algx
+# import bedlam.linked_lists
+# import bedlam.builder
+# import bedlam.build_matrix
+# import bedlam.coordinate_handler
+# import bedlam.shape_reader
