+++ bedlam/linked_lists.py
