+++ solutions/example_solutions.txt
@@ -1,2 +1,10 @@
 [0, 1, 3]
 [4, 2]
+[0, 1, 3]
+[4, 2]
+[0, 1, 3]
+[4, 2]
+[0, 1, 3]
+[4, 2]
+[0, 1, 3]
+[4, 2]
