--- tests/linked_list_add_tester.py
@@ -1,8 +1,0 @@
-from linked_lists import LinkedList, ColumnObject
-
-def main():
-    root = ColumnObject(name='root')
-    headers.add_horizontally(root)
-
-if __name__ == '__main__':
-    main()
